+++ tests/text_tox_timed_initialization.py
@@ -26,33 +26,19 @@ def test_timed_initialization():
     D.tdve["tx_rate"].ts[0].insert(2021, 0)
     D.tdve["tx_rate"].ts[0].insert(2022, 0.5)
 
-
-    def make_constant_parset(F, D, year):
-        D2 = sc.dcp(D)
-        for table in D2.tables():
-            for ts in table.ts.values():
-                ts.insert(None, ts.interpolate(year))
-                ts.t = []
-                ts.vals = []
-        return at.ParameterSet(F, D2, "constant")
-
-
     def set_initialization_basic(F, D, year, y_factor=True):
-        # ps1 = make_constant_parset(F, D, year)
-        res = P.run_sim(make_constant_parset(F, D, 2018))
-
+        # Conventional approach using Y-factor adjustment only
+        res = P.run_sim(P.parsets[0].make_constant(2018))
         ps2 = at.ParameterSet(F, D, "basic")
 
         # Set initial compartment sizes for initialization quantities
         for qty in itertools.chain(F.characs.index[F.characs['setup weight'] > 0], F.comps.index[F.comps['setup weight'] > 0]):
             for pop in D.pops:
-
                 if y_factor:
                     ps2.pars[qty].meta_y_factor = 1
                     ps2.pars[qty].y_factor[pop] = res.get_variable(qty, pop)[0].vals[-1]/ps2.pars[qty].ts[pop].interpolate(year)
                 else:
                     ps2.pars[qty].ts[pop].insert(year, res.get_variable(qty, pop)[0].vals[-1])
-
 
         return ps2